requirements.txt
python-telegram-bot>=20.0,<22.0
Pillow>=10.0.0
pytesseract>=0.3.10
boto3>=1.34.0
